chore: refactoring Entity Page
