+++ plugins/testkube/src/components/TWESummaryMetricsComponent/index.tsx
@@ -1,0 +1,1 @@
+export { TWESummaryMetricsComponent } from './TWESummaryMetricsComponent';
