+++ plugins/testkube/src/components/TWExecutionsDetailedTableComponent/index.tsx
@@ -1,0 +1,1 @@
+export { TWExecutionsDetailedTableComponent } from './TWExecutionsDetailedTableComponent';
